cross_modal_retrieval_system/backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "cross-modal-retrieval-backend"
version = "0.1.0"
description = "Backend service for cross-modal retrieval system."
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "confluent-kafka",
    "fastapi",
    "httpx",
    "numpy>=2.4.2",
    "pillow",
    "pydantic-settings",
    "pymilvus",
    "python-multipart",
    "torch",
    "torchvision",
    "uvicorn",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]

[tool.uv]
package = false
